Closes-bug: #1420035 Fix unit test code compilation issues in trusty
'scons test flaky-test' found few issues when ran in trusty 14.04.
Some tests still time out, but with these changes, code should least
compile.

+++ src/base/test/connection_info_test.cc
@@ -74,7 +74,7 @@ class ConnectionInfoTest : public ::testing::Test {
         const std::string ctype(
             g_process_info_constants.ConnectionTypeNames.find(
                 ConnectionType::TEST)->second);
-        for (int i = 0; i < vcinfo->size(); i++) {
+        for (size_t i = 0; i < vcinfo->size(); i++) {
             ConnectionInfo &tinfo(vcinfo->at(i));
             if (tinfo.get_type() == ctype &&
                 tinfo.get_name() == name) {
